requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
pyjwt==2.8.0
bcrypt==4.1.2
python-socketio==5.10.0
python-dotenv==1.0.0
pydantic==2.5.0
